QI.9v5kg2b: Refactorings
Starting to lift some traits out of the way-too-large SpaceHistory.
